refactor: Product 패키지 변경
